aggiunto controllo su installazione dei prerequisiti

+++ src/commands/calamares.ts
@@ -19,7 +19,7 @@ export default class Calamares extends Command {
   async run() {
     const { args, flags } = this.parse(Calamares)
 
-    if (Utils.isRoot()) {
+    if (Utils.isRoot() && Utils.prerequisitesInstalled()) {
       console.log(`>>> eggs: removing calamares...`)
       shx.exec('rm /etc/calamares -rf')
       shx.exec('apt-get remove --yes --purge calamares calamares-settings-debian')
